tests/cxx_telemetry.cpp: Remove it
We added that file in 9f4c74da (Complete opentelemetry addition from
previous commit, 2024-01-25) so that we can check in CI that the opentelemetry
dependency is available and we could compile code using it.

Now that we have proper telemetry code which uses the opentelemetry API
we don't need it anymore.