Attribute::fire_change_event: Create AttributeValue_4 for IDLv4 device
This change is "obvious" in the sense that without this change the `vers
== 4` branch cannot be taken.  However, I cannot seem to write a test
which demonstrates a problem when we don't have this change.

The only difference between AttributeValue_5 and AttributeValue_4 is the
presence of a data_type member for AttributeValue_5.  This seems to
allow overriding the data_type that is sent to the subscriber (from the
value in the AttrValUnion), but it is not clear to me when this is
needed.