cmake: Handle zlib dependency better
For telemetry we need zlib as dependency. But we never checked for it on
windows but instead passed in ZLIB_INCLUDE_DIR and ZLIB_LIBRARY.

Instead we now use find_package. For the installation and building the
Catch2Tests we need the zlib DLL so we create
ZLIB_RUNTIME_RELEASE/ZLIB_RUNTIME_DEBUG if not present.

This also fixes the issue that telemetry code could not be used in
Catch2Tests.