configure: Add compile test for expected opentelemetry variant type
The telemetry support in cppTango is developed against std::variant and
not the compatibility type in opentelemetry-cpp
opentelemetry::nostd::variant.

And as cppTango already requires C++17 the quickest fix is to check at
configure time that opentelemetry-cpp was compiled as we expect it.

This is also more future proof as opentelemetry-cpp already has dropped
support for C++11 [1].

[1]: https://github.com/open-telemetry/opentelemetry-cpp/issues/1830