tests/catch2/Context: Keep environment vector as class member
The idea of append_std_entries_to_env is that the returned std::string
vector owns the memory the env vector entries are pointing at.

But that requires that all entries from the passed std::vector<const
char*> are in std::vector<std::string>.

With keeping the return value of append_std_entries_to_env as class
member we ensure that the lifetime of the memory pointed to by env vector
exceeds the Context constructor.

Adding a trailing nullptr is now also handled in TestServer:start already.