tests/TestServer::start: Increase lifetime of argument and environment vector
Just to be sure we now keep the env and args vectors as class members
around. The reason is that we don't know when the implementations of
start_server access these members. Only if we keep them around until the
server is finished we can be sure that no invalid memory access happens.