zmq_event_subscription_change: Default ALARM_EVENT to idl6
on the server when reconnecting.

When a client tries to reconnect to a zmq event topic it invokes the
ZmqEventSubscriptionChange command with DevVersion "0".

Rather than special casing this on the client side for ALARM_EVENT (by
invoking ZmqEventSubscriptionChange with DevVersion "6"), we should
instead follow the pattern for other events and set the default
client_lib_version to 6 for ALARM_EVENT.