catch2_alarm_event.cpp: Test for event reconnection
This new test is slow as we have to wait for the heartbeat thread which
polls at 10 seconds.

You can skip the "[slow]" tests by running `Catch2Tests ~[slow]`.

To implement this test, I have added the ability to restart a TestServer
on the same port.  I don't particularly like how the relationship
between the TangoTest::Context and TangoTest::TestServer classes looks
now as I think it is a bit unclear what the responsibility of each class
is now.  However, in the interest of time, I am going to save the
rethink and refactor for later.