.gitlab-ci.yml:clang-tidy: Enable BUILD_TESTING=ON
As we want to check the catch2 tests we need to enable BUILD_TESTING=ON
so that the tests get added to the compile_commands.json generated by CMake.

This was missed when we added the catch2 tests to be checked by
clang-tidy in: 91d814c7 (.gitlab-ci.yml: Run clang-tidy against cacth2
tests, 2024-02-09).