telemetry.cpp: Avoid lifetime "issues" with static storage objects
The opentelemetry-cpp has a lot of helper functions like

static RuntimeContextStorage *GetDefaultStorage() noexcept;

which allow for convenient usage of global objects.

But during application shutdown we don't know that these static storage
objects are kept around long enough for us until all threads are finished.

So let's keep a shared_ptr copy of these objects around or in the case of
SpanContext::GetInvalid() create a new one everytime we need it.