Attribute::delete_seq: Set data_size = 0
As the sequence has been be freed at set to nullptr, we should also
indicate that the sequence is empty by setting the data_size to zero.

This gets us back to the same state as before we called `set_value`.

This avoids a segfault if we call check_alarm after push_event as in
catch2_alarm.cpp.