configure/cmake_linux.cmake: Do not install pkgconfig twice
We are already installing the pkgconfig file in configure/CMakeLists.txt
so we don't need to install it here too.  Also, this install is using
CMAKE_INSTALL_FULL_LIBDIR which does not work with

    `cmake --install build --prefix <x>`

We should install the pkgconfig file on Windows as cygwin users might be
able to make use of it.