DServerSignal::ThSig: Join the thread when we shutdown
ThSig accesses static objects so must be joined before we return from
main.

We are extending the existing SIGINT mechanism for sending commands from
the DServerSignal singleton to the thread so that we can request that
ThSig stops. We could probably have got away with just sending a SIGTERM
as we are almost certainly shutting down, but I feel adding this "stop
command" makes the intention clearer.