tango_catch2_tests.cmake: Copy both JPEG libraries
COMMAND_EXPAND_LISTS does not work with lists inside generator
expressions, which means that the copy command fails if both the DEBUG
and RELEASE versions of libjpeg are found.

The simplest way to fix this is to just copy both if we find both rather
than trying to be clever.

This has been broken since ed464248 (tango_catch2_tests.cmake: Copy DLLs
to test directory, 2024-03-11).