Catch2 tests: Call init_device implicitly
The C++ device server interface was designed so that people implementing
device servers have to call init_device and delete_device manually in the
derived constructor and destructor.

While this is a possible design and also what POGO does, most of the
test authors were not aware of that.

Adding a manual call to each constructor also requires people implementing
the constructor so in order to keep the test code short we call
init_device now in AutoDeviceClass::device_factory just after creating the
DS object.

We can't do the same for delete_device, so we document our way out.