ci/windows-build.ps1: Use Push/Pop-Location
In contrast to bash, the call operator '&' of powershell does not spawn
a new shell process to execute the script, but instead introduces a new
scope in the current process.

Unfortunately, the "Current-Location" is not part of a scope, but
instead some global state for the powershell process.  As such, if we
change directory in the script this change of directory happens in the
calling shell too.

Using Push/Pop-Location here allows us to ensure that the script doesn't
modify the global Current-Location.  This fixes the following error on
the CI:

    $ try { & '.\ci\windows-test.ps1' } catch { Write-Host $_; exit 1 }
    The term '.\ci\windows-test.ps1' is not recognized as a name of a cmdlet, function, script file, or executable program.
    Check the spelling of the name, or if a path was included, verify that the path is correct and try again.