TestServer: Introduce an ExitStatus type
On POSIX systems a process does not have an exit status if it was
terminated by a signal and our unix platform layer was incorrectly using
WEXITSTATUS in this case.  This always returns zero if the process was
terminated by a signal so the test infrastructure does not notice
anything is wrong.

By using a rich ExitStatus type the platform layer can communicate with
the test infrastructure that the process has been aborted in a platform
independent way.