TestServer: Do not call stop() on start() timeout
In a real test if starting the server has timed out then we are about to
unwind the stack anyway, so the dtor will be called where we call stop()
making calling stop() in start() unnecessary.

However, calling stop() in start() hinders the test_test_server test
which checks for timeouts being reported correctly because we have logs
generated the stop() call interfering with the test.  This is an issue
now that we correctly identify that the TestServer is terminate due to
an uncaught exception when we send SIGTERM before the DServer has been
fulling initialised.