catch2/utils/test_test_server.cpp: Always flush cout
The tests are relying on these strings being present.  On some platforms
(e.g. Linux) we can rely on the buffer being flushed for us, however,
the only platform agnostic way to ensure the strings are present in the
file when we expect them to be is to flush the stream ourselves.