tests/catch2/utils: Always print the filename prefix
When we run with --log-file-per-test-case we do not output the name
of the log file at the start of the test run (because there are
potentially multiple log files).

As we use this option for ctest, it can be difficult to track down
which file this corresponds to the failure you have just seen when you
have run the tests via ctest multiple times.  Adding the filename
prefix to the test output makes it easier to find the log files
in case of a failure.