tests/catch2: Do not require second initial event
when first subscribing.

As is explained in the comment in catch2_alarm_event.cpp there is a
race condition which means that this second event is not always present.

As the fact that we get these two events is a bit of an anomaly I don't
think it is worth spending time ensuring that they are always both
present.

When investigating this I thought of a scenario where I think this race
condition might cause a real problem, however, I think it is unlikely to
occur so I have not worked out a fix here.  See #1301.