TANGO_INSTALL_DEPENDENCIES: Only install required files
Before this commit, we were installing "whatever we could find" and often this
would include libraries which aren't actually needed.  Using the dependencies
of the tango target as a source of truth also avoids duplicating the logic
about when certain dependencies are needed.

With this commit we no longer install the cmake config files for the
telemetry dependencies, this will be address in a later commit.