TANGO_INSTALL_DEPENDENCIES: Use a special TangoConfig.cmake
... which hardcodes our dependency targets.

When we install our dependencies we know exactly which libraries to link
against as we are providing them.  Moreover, we don't want cmake to search
for the libraries again when compiling e.g. a device server because
we want to use _only_ the bundled dependencies.

By generating cmake logic to reconstitute the dependencies we used to build
the library we ensure the end user will link against these dependencies.