FindJPEG.cmake: Add our own and use it
This version of FindJPEG fits in with the reset of the find modules we
use, it is based on FindZeroMQ.cmake.  It has the following improvements:

- Provides a JPEG::JPEG-static target so that we can link to this on
Windows when building a static library
- The JPEG::JPEG target is SHARED IMPORTED when appropriate meaning that
the TARGET_RUNTIME_DLLS generator expression will include the JPEG DLL
for targets which link against it, e.g. Catch2Tests

The find module does not locate the turbojpeg libraries which were
by installed on Windows.  These are not required as the jpeg library
does not depend on them.  For example:

    > dumpbin.exe /dependents jpeg62.dll
    Microsoft (R) COFF/PE Dumper Version 14.29.30151.0
    Copyright (C) Microsoft Corporation.  All rights reserved.
    Dump of file jpeg62.dll

    File Type: DLL

      Image has the following dependencies:

        VCRUNTIME140.dll
        api-ms-win-crt-heap-l1-1-0.dll
        api-ms-win-crt-stdio-l1-1-0.dll
        api-ms-win-crt-runtime-l1-1-0.dll
        api-ms-win-crt-environment-l1-1-0.dll
        KERNEL32.dll

      Summary

            1000 .data
            3000 .pdata
           38000 .rdata
            1000 .reloc
            1000 .rsrc
           4A000 .text