Attribute::check_alarm(): Always log_quality() before return
In the previous two commits we stopped calling `log_quality()` if
the quality is ATTR_ALARM or ATTR_WARNING.  This is to avoid duplicate
logs if `check_alarm()` is called multiple times between read attribute
calls.

This is important if a user read attribute function has called
e.g. `set_value_date_quality(..., ..., ATTR_ALARM)`.  We should generate
a log in this case and so we need to call `log_quality()`.

Unfortunately, this means that we still generate multiple logs in
`check_alarm()` is called multiple times between calls to `set_value()`.
However, I think this is the lesser of two evils.