Catch2Tests: Increase the default polling period on macOS CI
For some reason on the macOS CI we sometimes get the polling thread
stalling for ~200ms.  When this happens, it can cause multiple polling
work items to be missed resulting in test failures.  I assume this is
something to do with the gitlab saas runners being heavily loaded.

Increasing the polling period just on the macOS CI to avoid this problem
as it doesn't seem to occur elsewhere.  The normal 100 ms polling period
is an artificially short to have the tests run quicker, so it is not a
big deal that this doesn't work with such a polling rate on a macOS VM
under load.