fire_<x>_event: Clear quality and alarms when deleting seq
If we manually push an event with e.g. push_change_event then we will
compute the alarm state of the attribute _if and only if_ someone has
subscribed to events from that attribute.

This means that if we call check_alarm after push_change_event we get a
different answer depending on whether anyone has subscribed.  As the
value that gets set by push_change_event gets cleared anyway, it is most
straight forward to also clear the quality/alarm value when we clear the
value.

In this way, it is as if `push_<x>_event` resets the attribute to a
state where it has no value, as if it has just been constructed.