Telemetry: Centralize configuration management
We now do all telemetry configuration in telemetry::Configuration() to
ensure that we have a valid configuration once the object is constructed.

We also turn Configuration into a class and make many more methods/members
private.

Only with these changes the added test passes, as the test requires that
we check all environment variables before trying to construct a telemetry
interface. So e.g. we now report with an exception that endpoints are
invalid even if telemetry itself is disabled, which was not the case
before.