DeviceImpl::dev_state(): Do not cleanup previous attr on except
In c6782dca (- Fix bug 749 (user dev_state method not calle when att
throws exception + hide exception in state method), 2015-10-26), we
stopped re-throwing attribute read exceptions that occurred during a
`dev_state()` called, however, we left the cleanup code which called
`Attribute::delete_seq()` on all the previously read attributes.

This causes a discrepancy between the behaviour of `dev_state()` if an
attribute has an alarm and is read before the exception is throw
compared to if an attribute has an alarm after the exception is thrown.

This commit ensures that the order the attributes are read does not
matter in this regard.