configure/CMakeLists.txt: Workaround cmake limitations
When superprojects like the TangoSourceDistribution include cppTango via
add_subdirectory these super projects need to be able to fake successfull
find_package calls for dependencies like tangoidl.

But that is only possible if the package is not required. We can still
error out if it was not found as tangoidl_FOUND can be overriden by the
superproject.