DeviceImpl::dev_state: Always set attribute flags before read
Before this commit, we were only setting the attribute value flag and
wanted_date flag before the read for >IDLv3 devices, however, we were
always using these flags to workout how to cleanup in case of an
exception.

Now we correctly cleanup, regardless of which IDL version the server is
using.