TANGO_INSTALL_DEPENDENCIES: Install dependencies for all configurations
Some imported dependencies might not have a IMPORTED_CONFIGURATIONS for
the configuration selected by the user at build time.  In this case
CMake will use the first configuration listed in IMPORTED_CONFIGURATIONS
as a fallback.

When installing the dependencies, we need to install the
IMPORTED_LOCATION that is actually used for the configuration selected
which means possibly the fallback configuration.

In this commit we ensure we always install something for each dependency
regardless of configuration and that the IMPORTED_CONFIGURATIONS is
constructed in the same order in the CMake config file we install.

This has been broken since c3644b7b (TANGO_INSTALL_DEPENDENCIES: Only
install required files, 2024-07-04) but was hidden because we did not
build on the Windows CI with Debug and Otel previously.