configure/CMakeLists.txt: Fix otel exporter name try_compile test with MSVC
In 72e2d7b67 (configure: Prefer compile tests for version dependent
features, 2024-02-29) we added a try_compile test to handle differences in
opentelemetry class naming.

Only now as we are updating the opentelemetry package to version 1.16.0
the try_compile test actually needs to pass as only the new class is
available.

The documentation for try_compile mentions CMAKE_TRY_COMPILE_CONFIGURATION
which needs to be set so that try_compile uses the correct configuration.
This directly influences CMAKE_MSVC_RUNTIME_LIBRARY which depends on it.

By setting CMAKE_TRY_COMPILE_CONFIGURATION to the CMAKE_BUILD_TYPE we do
fix the issue, although making it even more mandatory to be set.

See also [2] for a minimal repro.

The fix here also breaks omniORB try_run invocation as we hardcoded the debug build type there.

[1]: https://cmake.org/cmake/help/latest/command/try_compile.html#other-behavior-settings
[1]: https://discourse.cmake.org/t/try-compile-and-genex-in-cmake-msvc-runtime-library/11757/3