tests/catch2: Revise matchers
As we now have the STL CORBA helpers, we can implement more high-level
matchers and use them. This not only enhances the error messages when a
a test fails, but also allows us to prefer types other strings for e.g. in
the case of Tango::ALARM_EVENT.

The AnyErrorMatches matcher was also removed in favour of
EventErrorMatches plus one of the builtin range matcher.