cmake: Add Threads::Threads dependency to cppTango
We always required thread support, but currently only search for it in the
tests (historical reasons).

This setup currently breaks compilation with clang 19 and PCH.

Fix it by moving the dependency up to the tango project. That also
requires that we now require that package in the cmake config files. And
for windows this means we need to special case our dependency handling.