tests: Move matchers to namespace
We want to define additional matchers to ease testing. But this makes
using them currently difficult as we have to pull in every symbol.

By putting them into a namespace and including the matchers namespace
locally we can avoid too much boilerplate code while avoiding to include
the matchers gobally.