tests: Add StringMaker template specialications
This gives us better fail messages. This is only a start though.

Defining CATCH_CONFIG_ENABLE_ALL_STRINGMAKERS is not the cleanest solution
but allows us to reuse already defined stringmakers from catch2.

We also opted for having the member names in the output as this output is
targeted for developers. The same reasoning applies for having it compact
in a JSON-like format.