tests/.clang-tidy: Disable some checks
In catch2 tests it is common to write

REQUIRE(a == b);

but that triggers the following error:

/home/thomas/devel/cppTango/tests/catch2_cmd_polling.cpp|47 col 9| error: chained comparison 'v0 <= v1 == v2' may generate unintended results, use parentheses to specify order of evaluation or a logical operator to separate comparison expressions [bugprone-chained-comparison,-warnings-as-errors]
||    47 |         REQUIRE(idlver == device->get_idl_version());

We have to assume that catch2 does it correctly, so let's disable the
check bugprone-chained-comparsion for the tests.

We also don't need to care about limiting external linkage and don't care
about unchecked optional access.