detail::has_corba_extract_operator_to: Fix it
We need to add an lvalue reference to T so that this matches. As we have
never used CORBA::Any for our matchers accepting AnyLike objects, this was
not found earlier. Add a static_assertion as well.

Bug present since the initial version in 8a3f41fd (catch2: Add matcher for
any-like types such as DeviceData, 2024-02-13).