tests: Fix obvious memory leaks
This fixes the very obvious memory "leaks" in the tests. As ASAN does not
distinguish between these "leaks" and real memory leaks in cppTango
itself, this makes the CI ASAN test output less noisy already.

We also opted for the quickest fix, as the old and cxx tests are not
planned to be maintained/changed anymore.