.clang-tidy: Enable cppcoreguidelines-virtual-class-destructor
See [1] for an in-depth explanation. As it turns out this has not found
any issue in the production code, only one in the tests.

It is also not clear why clang-tidy complains about protected virtual
destructors. This seems to be correct in forward thinking because derived
classes will inherit the virtual-ness and thus be correct.

[1]: https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rc-dtor-virtual