Attribute: Clear allocated memory
In the Attribute destructor we need to clear our allocated memory in the
Attr_Value union.

But for that to work we need to ensure that we start with a nullptr in it
and also only call delete_seq as that also nulls the pointer.

This fixes a memory leak.