telemetry.cpp: Make it more compatible with different opentelemetry compilation options
opentelemetry-cpp has the option WITH_STL which allows to tune the usage
of STL classes vs the builtin ones. It defaults to OFF.

Until now we never passed the WITH_STL cmake option to compile opentelemetry-cpp.

But starting from v1.17.0 this breaks cppTango as the provider factory
methods now return std::unique_ptr<XXXProvider> but the
SetLoggerProvider/SetTracerProvider methods accept only a
opentelemetry::nostd::shared_ptr<XXXProvider>.

The breakage can be circumvented if opentelemetry-cpp is compiled with
WITH_STL=CXX17 or higher as then opentelemetry::nostd::shared_ptr is an alias to
std::shared_ptr. And this has a constructor accepting a std::unique_ptr,
whereas the nostd type has not.

But it is tedious/impossible to teach all package managers out there the
"correct" way to compile opentelemetry-cpp, so let's use a wrapper
template function to accomodate compilation of opentelemetry-cpp with and without
the WITH_STL cmake option.