src: Generalize STL helpers for CORBA Sequence classes
As it turns out the STL helper functions added in ff49ea9f (tests/catch2:
Add STL helper functions for CORBA types, 2024-09-24) don't work with the
string array classes. The reason is that these don't inherit from
_CORBA_Sequence.

The fix is to generalize is_corba_seq_v, as it now checks for a static
allocbuf function, and then use that with SFINAE in the free functions.
The return values also have to change as the string array classes don't
have the NP_data member.

As get_buffer allocates memory when asked to return the buffer for empty
sequences so we handle that case separately.