Refactor condition struct into event class
The condition structure held a condition variable, a mutex and a flag --
what's commonly called an "Event" in Python. Like the Python class, it
was used to set the flag, and wait on the other end for the flag to be
set. These operations in Python's Event class are called set() and
wait().

This commit refactors the condition struct, turning it into the event
class, and moving the code that was effectively performing the set() and
wait() operations into new named methods of the class. This makes the
code easier to reason about, and easier to reuse for other purposes.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>