Improve kill/waitpid of child process
The test code currently issues a SIGTERM signal to the child process
running the device server, and waitpid()'s on it. The call to waitpid is
done without blocking, and the code assumes the child process will have
finished by then (the test checks that the child process doesn't exist
after it's been waitpid'd, and fails if it's still alive). To make this
more likely, the test waits for 1 second in between the kill and waitpid
operations. This is of course both brittle (it might not be enough time
in some cases), and sub-optimal (most times it probably takes less than
that).

This commit restructures the wait the child process is waited after the
SIGTERM signal is sent. Instead of waiting for a full second and then
taking action, the code now issues non-hanging waitpid() calls in a
loop, checking whether the child process has exited or not, periodically
retrying, with a 100ms sleep between retries. If after a maximum waiting
time of 5 seconds the child process is still alive, SIGKILL is sent and
the test fails. With these changes, the kill+waitpid logic now takes a
fraction of a second rather than more than a second.

While working on this I noticed that the reporting of the exit status
was in need of some minor fixes, mainly to show in the report only the
bits of information that make sense to show given how the child process
exited, #ifdef-ing the usage of WCOREDUMP, plus some minor cosmetic
changes in the code that generates the report.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>