tests/multi_thread_sighandle: Simplify command line argument type
On the one hand it's easier if we declare this array as const char *, since all
its members are built-in constants (so we don't require explicit casts
on each individual member -- and *C*-style casts for that matter); then
when we pass these arguments to the relevant Tango API we can const_cast
them.

Secondly, let the compiler figure out the size of the array instead of
declaring it ourselves.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>