DServerSignal: Replace sigwait with thread-safe signal handlers
Current state
-------------

Tango is itself interested in being notified about some signals (e.g.,
SIGINT), called "auto" signals in the code. User devices/classes can
express interest in being notified about any signal, either one of the
"auto" signals or any other arbitrary one.

The current way in which Tango deals with signals in general is based on
a synchronous/blocking signal handling strategy. First, all signals in
the main thread are blocked, and signal handlers for the "auto" signals
are reset to SIG_DFL to avoid running handlers users might have
registered previously. Then a "signal thread" is spawned that unblocks
the "auto" signals, and continuously loops calling sigwait(). The
rationale of this technique is that since the main thread (and the other
threads spawned by Tango itself after that) have all signals blocked,
then it is safe to assume that the signal thread will receive all
signals sent to the process by the OS.

On top of handling these auto signals, users can register callbacks for
arbitrary signals that can be executed in two modes. In the first, the
callback is executed in the signal handler thread. To achieve this, the
requested signal number is first recorded in memory, and then the signal
thread's loop is triggered via pthread_kill(SIGINT) to unblock the
requested signal number in future sigwait() calls, thus allowing those
signals to be delivered to users. In the second mode, the callback is
executed directly in the context of a OS signal handler; for this to
work, an actual handler needs to be installed separately from the signal
thread, and the signal thread must *not* be requested to process these
signals. These callbacks can then be unregistered, which again,
depending on the mode of operation, requires communicating with the
signal thread, or installing the default signal handler for the signal.

Problem
-------

Although in principle all of this should work, there is a caveat: users
might spawn *other* threads themselves. If created before Tango is
initialised, these other threads' signal masks, more often than not,
will allow the OS to deliver signals to them, hence breaking the
assumption Tango does about being able to catch all signals with
sigwait() in the signal thread. If these user-created threads are
created after Tango is initialised, there is also no guarantee that
their signal mask will block all incoming signals. In any case, whether
user threads are created before or after Tango initialisation, it
impossible for Tango to exert any control over them and alter their
signal masks.

Note that this issue of user-created threads has been observed time and
time again in pytango. One of the prevalent reasons is that the OpenBLAS
library bundled with the pre-compiled numpy wheels one gets from PyPI
for Linux installations spawn threads at import time, so unless one sets
the OMP_NUM_THREADS=1 environment variable one will hit this issue.

This work
---------

This commit fixes this situation by reworking how signals are handled by
Tango altogether. Instead of relying on threads to block signals, and
letting the signal thread wait with sigwait(), this new approach is
based on setting signal handlers that can be run by any thread. No
special signal mask is set on any thread. When a signal needs to be
received, a "enqueueing" signal handler is set for it that puts the
signal number in a SynchronisedQueue object (see previous commit). The
signal thread in turn pops these signals from the queue and reacts as
usual. To ensure that not two signal handlers are running at the same
time, the enqueueing signal handler is set with a full sa_mask, meaning
that all other signals are blocked for the duration of the signal
handler. Also to avoid potential problems where the enqueueing signal
handler could be invoked exactly while SynchronisedQueue::get() is
holding the underlying locked mutex (and therefore the signal handler
would raise an exception), we block all signals from reaching the signal
thread.

In this new scheme, when users register callbacks for arbitrary signals,
this can now be done directly by setting the enqueueing signal handler
for those signals. Previously this involved communicating with the
signal handler thread, and therefore the resulting code is now simpler
in that regard. For the mode where users register callbacks to be
invoked directly in the context of a signal handler, a "direct" signal
handler is set. When a certain signal does not need to be handled, the
default signal action is set back regardless of which mode the user
previously requested. Again, this is a much simpler scheme than it
previously was, leading to the removal of some synchronisation between
DServerSignal and ThSig, and thus to the removal of a few fields in both
of them.

Notes
-----

Signal handlers are not installed in the DServerSignal constructor
anymore, but in the initialise() method instead. This is because the
signal handlers access the DServerSignal::instance(), which at
initialise() time will be available.

In order to keep the code a bit tidier, I also unified the installation
of signal handlers, including always throwing an exception when a signal
handler failed to be installed. This was being done in all calling
places except when installing the default handler for the auto signals
in non-Window platforms (but in Windows an exception was thrown in the
case of failures), where a warning was printed to stderr. This asymmetry
felt incorrect, so I decided to raise an exception in these cases as
well.

Finally, and as the icing on the cake, the code that tested the
behaviour of signal handling in Tango now can safely assert that sending
SIGINT/SIGTERM to a Tango device server running without *or with* a
user-provided thread always exits with status 0.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>