DServerSignal: Initialise static data statically
While it is not invalid or incorrect to initialise the sign_name static
member within DServerSignal's constructor's, it's better to be clear
about this happening in a static context. An even better solution of
course would be to be able to constexpr initialise these, but that's a
harder problem to solve.

While tinkering with this piece of code, I also took the opportunity to
simplify the int->str conversion.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>