DServerSignal: Start signal thread separately from DServerSignal ctor
The signal thread receives a pointer to the DServerSignal instance,
which at this point is not formally fully initialised, and therefore it
is not safe to use. Moreover, once we refactor the code to use signal
handlers, we will also want to make sure the DServerSignal instance is
fully initialised before the signal handler can reference it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>