DServerSignal: Unify Windows and non-Windows signal handling
In Windows platforms, signals were always caught using a signal handler,
stored in memory, and then communicated to the signal thread via a
Windows-specific "Event" object. We are now doing basically the same
thing in non-Windows platforms (although a bit more safely, as we
properly lock around the shared data structure used to communicate
between threads). It thus makes sense to unify the implementations.

This commit removes the Windows-specific logic, reusing the same logic
used in non-Windows platforms to record incoming signals that need to be
dispatched by the signal thread. Like before, the resulting code is
easier to maintain and follow.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>