DServerSignal: Put special value to request signal thread termination
The signal thread previously relied on receiving a SIGINT signal and
checking for a boolean flag set elsewhere to verify whether the
DServerSignal destructor had requested a shutdown. Since we have a
direct line of communication now, we can put a special value (that isn't
a valid signal number) to more explicitly, and less confusingly, make
the same request.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>