tests/catch2: Implement signal sending to test server
While all platforms support signals to an extent, the support in Windows
is limited, since there is not really a concept of sending signals, and
instead one is left with calling different APIs (that however result in
signal handlers being invoked in the target process). This has two main
implications:

 * The logic to send a signal is platform-specific, and
 * The set of signals we can send (and in particular, the signals that
   are relevant to a Tango device server) is also platform-specific.

This commit introduces the implementation of two new methods at the
TangoTest::platform level that provide for those two points. On top of
that, the TestServer class offers a convenient wrappers for both.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>