tests/catch2/test_server: fix treatment of wait_for_exit() result
The TestServer::wait_for_exit() method, as its name implies, waits for
the server process to exit, without reporting on the exit status,
deferring that to the stop() method that occurs later on (either
explicitly, or via the dctor). This is done by storing the exit status
in a std::optional, which stop() inspects before attempting to actually
stop the server process.

There seemed to be an issue with how this was done though: in principle,
if a wait_for_exit result is present, stop() should use that as-is,
otherwise it should attempt to stop the server or detect an early,
unexpected exit. This is not what was happening: an error in the logic
blocks meant that even if there was a wait_for_exit result present,
stop() still tried to stop the server or detect an early, unexpected
exit. This resulted in the exit result kind being "EarlyExitUnexpected",
which for an otherwise healthy server exit generated a warning in the
logs.

This commit makes sure the stop-or-detect-early-exit logic is not
executed when a wait_for_exit result is already present.

This has been broken since 6c7c9cbf (TestServer::wait_for_exit: Add it,
2024-06-14).