tests/catch2_server: Fix and reincorporate test
Since this test was introduced in 1b65fb73 (tests/catch2_server.cpp: Add
test for DServer Kill command, 2024-06-14) it has never been included in
the test suite.  There was probably some issue with rebasing the MR
which introduced it.  If this test had been properly included, it would
have caught the issue from the previous commit and also the fact that
the API is being used wrongly here.

It can now easily be added into the suite, after fixing a small API rot.