tests/catch2: Start optional dummy thread before Tango starts
Before the tango system is started by the TestServer entrypoint, users
can now indicate via a environment variable that they want a new thread
to run before the Tango system is initialised. This mimics what happens
in pytango (and possibly other scenarios), with threads pre-existing
Tango, which in turn used to confuse the signal handling logic in Tango.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>